% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{betas.Gy}
\alias{betas.Gy}
\title{Example beta list of hJAM}
\format{An object of class \code{numeric} of length 19.}
\usage{
betas.Gy
}
\description{
The simulated beta list for the examples in hJAM package
}
\keyword{datasets}
