% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conditional_A}
\alias{conditional_A}
\title{Example conditional A matrix of hJAM}
\format{An object of class \code{matrix} with 19 rows and 2 columns.}
\usage{
conditional_A
}
\description{
The conditional A matrix for the examples in hJAM package
}
\keyword{datasets}
