% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conditional_Z}
\alias{conditional_Z}
\title{Example conditional Z matrix of hJAM}
\format{An object of class \code{matrix} with 19 rows and 2 columns.}
\usage{
conditional_Z
}
\description{
The conditional Z matrix for the examples in hJAM package
}
\keyword{datasets}
