% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EN.hJAM.R
\name{EN.hJAM}
\alias{EN.hJAM}
\title{Elastic net hJAM}
\usage{
EN.hJAM(
  betas.Gy,
  N.Gy,
  raf.Gy = NULL,
  Geno,
  A,
  tune_glmnet = 0.5,
  ridgeTerm = FALSE
)
}
\arguments{
\item{betas.Gy}{The betas in the paper: the marginal effects of SNPs on the phenotype (Gy)}

\item{N.Gy}{The sample size of the GWAS where you obtain the betas.Gy and betas_se.Gy}

\item{raf.Gy}{The reference allele frequency of the SNPs in betas.Gy}

\item{Geno}{The individual level data of the reference panel. Must have the same order of SNPs as in the betas.Gy.}

\item{A}{The conditional A matrix.}

\item{tune_glmnet}{The \eqn{\alpha} used in the glmnet R package to tune the shrinkage parameter. Default is 0.5.}

\item{ridgeTerm}{Add a small elelment to the diagnoal of X'X to make the matrix invertable.}
}
\value{
An object of the Regularized hJAM

\describe{
   \item{numSNP}{The number of SNPs that the user use in the instrument set.}
   \item{Selected_variable_length}{The number of selected intermediates, regardless of the credible sets.}
   \item{Selected_variable_name}{The label/name for each selected intermediates.}
   \item{Coefficients}{The coefficients of selected intermediates. Otherwise will be zero.}
}
}
\description{
Function to implement regularized hJAM, including elastic net hJAM and lasso hJAM.
}
\examples{
data(ENhJAM.SimulationSet)
EN.hJAM(betas.Gy = Simulation.betas.gwas, N.Gy = 5000, raf.Gy = Simulation.maf.gwas,
Geno = Simulation.Geno, A = Simulation.Amatrix, ridgeTerm = FALSE)
}
\author{
Lai Jiang
}
