% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ENhJAM.SimulationSet}
\alias{ENhJAM.SimulationSet}
\alias{Simulation.Amatrix}
\alias{Simulation.Geno}
\alias{Simulation.betas.gwas}
\alias{Simulation.betas.se.gwas}
\alias{Simulation.maf.gwas}
\title{Simulation data for EN-hJAM}
\format{
The \code{ENhJAM.SimulationSet} is a set of simulation data sets for the example of elastic net hJAM
\describe{
 \item{Simulation.Amatrix}{The conditional \eqn{\hat{A}} matrix with 118 metabolites and 144 SNPs, which was composed by SuSiE JAM and the marginal \eqn{\hat{A}} matrix.}
 \item{Simulation.Geno}{The reference genotype data for the 144 SNPs from the European-ancestry population in 1000 Genome Project (Consortium, 2015).}
 \item{Simulation.betas.gwas}{The b vector. The association estimates between selected SNPs and the risk of prostate cancer from (Schumacher et al., 2018)}
 \item{Simulation.betas.se.gwas}{The se(b) vector from (Schumacher et al., 2018)}
 \item{Simulation.maf.gwas}{The vector of the effect allele frequency of the SNPs from (Schumacher et al., 2018)}
}
}
\description{
Simulation data for EN-hJAM
}
