% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cond_Z.R
\name{get_cond_Z}
\alias{get_cond_Z}
\title{Compute conditional Z matrix}
\usage{
get_cond_Z(marginal_Z, Gl, N.Gx, ridgeTerm = T)
}
\arguments{
\item{marginal_Z}{the marginal effects of SNPs on the exposures (Gx).}

\item{Gl}{the reference panel (Gl), such as 1000 Genome}

\item{N.Gx}{the sample size of each Gx. It can be a scalar or a vector. If there are multiple X's from different Gx, it should be a vector including the sample size of each Gx. If all alphas are from the same Gx, it could be a scalar.}

\item{ridgeTerm}{ridgeTerm = T when the matrix L is singular. Matrix L is obtained from the cholesky decomposition of G0'G0. Default as True.}
}
\description{
The get_condZ function is to get the conditional Z matrix by using marginal Z matrix
}
\examples{
data(reference_data)
data(betas.Gy)
data(marginal_Z)
get_cond_Z(marginal_Z = marginal_Z, Gl = Gl, N.Gx = 1000, ridgeTerm = TRUE)
}
