% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mJAM_helper.R
\name{get_XtX}
\alias{get_XtX}
\title{Get transformed statistics: XtX}
\usage{
get_XtX(N_outcome, Gl, maf)
}
\arguments{
\item{N_outcome}{Sample size in the GWAS where we obtained `betas`}

\item{Gl}{A matrix of reference dosage, columns are SNPs and rows are individuals.}

\item{maf}{A vector of minor allele frequencies}
}
\value{
a variance covariance matrix of scaled Gl
}
\description{
To calculate sufficient statistics based on summary statistics
}
