% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mJAM_helper.R, R/mJAM_helper_chrX.R
\name{get_yty}
\alias{get_yty}
\title{Get transformed statistics: yty}
\usage{
get_yty(maf, N_outcome, betas, betas.se)

get_yty(maf, N_outcome, betas, betas.se)
}
\arguments{
\item{maf}{A vector of minor allele frequencies}

\item{N_outcome}{Sample size in the GWAS where we obtained `betas`}

\item{betas}{A vector of marginal estimates of effect sizes (betas for continuous outcome; logOR for binary outcome)}

\item{betas.se}{A vector of the standard errors of marginal effect estimates (`betas`).}
}
\value{
median of yty estimates across all SNPs; and a vector of all yty estimates

median of yty estimates across all SNPs; and a vector of all yty estimates
}
\description{
To calculate sufficient statistics based on summary statistics.
This yty estimate follows Yang et al. (2012) Nat Gen.
Marginal estimates from one SNP will produce one yty estimates.
Yang suggests taking the median across all SNPs to obtain a robust estimate.
Here we record all yty estimates and output both the median and the entire vector.

To calculate sufficient statistics based on summary statistics.
This yty estimate follows Yang et al. (2012) Nat Gen.
Marginal estimates from one SNP will produce one yty estimates.
Yang suggests taking the median across all SNPs to obtain a robust estimate.
Here we record all yty estimates and output both the median and the entire vector.
}
