% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mJAM_helper.R, R/mJAM_helper_chrX.R
\name{get_z}
\alias{get_z}
\title{Get transformed statistics: z, or Xty}
\usage{
get_z(maf, betas, N_outcome)

get_z(maf, betas, N_outcome)
}
\arguments{
\item{maf}{A vector of minor allele frequencies}

\item{betas}{A vector of marginal estimates of effect sizes (betas for continuous outcome; logOR for binary outcome)}

\item{N_outcome}{Sample size in the GWAS where we obtained `betas`}
}
\value{
a numeric vector of calculated z statistic

a numeric vector of calculated z statistic
}
\description{
To calculate sufficient statistics based on summary statistics

To calculate sufficient statistics based on summary statistics
}
