% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Gl}
\alias{Gl}
\title{Example reference data of hJAM}
\format{The \code{Gl} object is a data matrix with 500 individual of 19 simulated genotypes.}
\usage{
Gl
}
\description{
The simulated data for the examples in hJAM package
}
\keyword{datasets}
