% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{GTEx.PrCa}
\alias{GTEx.PrCa}
\alias{GTEx.PrCa.IVWmarginal.A}
\alias{GTEx.PrCa.marginal.A}
\alias{GTEx.PrCa.marginal.A.se}
\alias{GTEx.PrCa.inclusion.indicator}
\alias{GTEx.PrCa.Amatrix}
\alias{GTEx.PrCa.Geno}
\alias{GTEx.PrCa.betas.gwas}
\alias{GTEx.PrCa.betas.se.gwas}
\alias{GTEx.PrCa.pvalues.gwas}
\alias{GTEx.PrCa.maf.gwas}
\alias{GTEx.PrCa.rsid}
\alias{GTEx.PrCa.marginal.selected}
\title{Real data for selecting the genes on chromosome 10 for the risk of prostate cancer}
\format{
The \code{GTEx.PrCa} is a set of data sets which was applied for selecting the genes on chromosome 10 for the risk of prostate cancer
\describe{
 \item{GTEx.PrCa.IVWmarginal.A}{The marginal \eqn{\hat{A}} matrix with 158 genes and 182 eQTLs. The raw data was downloaded from GTEx analysis v7 (https://gtexportal.org/home/datasets). Priority Pruner was used to select the independent eQTLs. We used this matrix for MR-BMA implementation.)}
 \item{GTEx.PrCa.marginal.A}{The marginal \eqn{\hat{A}} matrix with 167 genes and 447 eQTLs. The raw data was downloaded from GTEx analysis v7 (https://gtexportal.org/home/datasets). This is the raw \eqn{\hat{A}} matrix for constructing the conditional weight matrix for SHA-JAM analysis.}
 \item{GTEx.PrCa.marginal.A.se}{The standard errors of the marginal \eqn{\hat{A}} effects for the SNP-gene pairs (167 genes, 447 eQTLs). The raw data was downloaded from GTEx analysis v7 (https://gtexportal.org/home/datasets).}
 \item{GTEx.PrCa.inclusion.indicator}{The inclusion indicator for the significant SNP-gene pairs (167 genes, 447 eQTLs). Significant as 1; otherwise 0. This matrix is for composing the conditional weight matrix using the raw data.}
 \item{GTEx.PrCa.Amatrix}{The conditional \eqn{\hat{A}} matrix with 167 genes and 447 eQTLs, which was composed by SuSiE JAM and the raw data of \eqn{\hat{A}} matrix.}
 \item{GTEx.PrCa.Geno}{The reference genotype data for the 447 eQTLs from the European-ancestry population in 1000 Genome Project (Consortium, 2015)}
 \item{GTEx.PrCa.betas.gwas}{The b vector. The association estimates between eQTLs and the risk of prostate cancer from (Schumacher et al., 2018)}
 \item{GTEx.PrCa.betas.se.gwas}{The se(b) vector from (Schumacher et al., 2018)}
 \item{GTEx.PrCa.pvalue.gwas}{The pvalues vector of the association estimates between selected SNPs and the risk of prostate cancer from (Schumacher et al., 2018)}
 \item{GTEx.PrCa.maf.gwas}{The vector of the effect allele frequency of the SNPs from (Schumacher et al., 2018)}
}
}
\description{
Real data for selecting the genes on chromosome 10 for the risk of prostate cancer
}
\references{
Consortium GP. A global reference for human genetic variation. Nature 2015; 526: 68.

Lonsdale, John, et al. The genotype-tissue expression (GTEx) project. Nature genetics 45.6 (2013): 580-585.

Schumacher, Fredrick R., et al. Association analyses of more than 140,000 men identify 63 new prostate cancer susceptibility loci. Nature genetics 50.7 (2018): 928-936.
}
