% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hJAM.R
\name{hJAM}
\alias{hJAM}
\title{Fit hJAM}
\usage{
hJAM(betas.Gy, N.Gy, Gl, Z, ridgeTerm = FALSE)
}
\arguments{
\item{betas.Gy}{the betas in the paper: the marginal effects of SNPs on the phenotype (Gy)}

\item{N.Gy}{the sample size of Gy}

\item{Gl}{the reference panel (Gl), such as 1000 Genome}

\item{Z}{the Z matrix in the paper: the marginal/conditional effects of SNPs on the exposures (Gx)}

\item{ridgeTerm}{ridgeTerm = TRUE when the matrix L is singular. Matrix L is obtained from the cholesky decomposition of G0'G0. Default as FALSE}
}
\description{
The hJAM function is to get the results from the hJAM model using input data
}
\examples{
data(reference_data)
data(betas.Gy)
data(conditional_Z)
hJAM(betas.Gy = betas.Gy, Gl = Gl, N.Gy = 5000, Z = conditional_Z, ridgeTerm = FALSE)
}
\author{
Lai Jiang
}
