% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hJAM_egger.R
\name{hJAM_egger}
\alias{hJAM_egger}
\title{Fit hJAM}
\usage{
hJAM_egger(betas.Gy, N.Gy, Gl, Z, ridgeTerm = FALSE)
}
\arguments{
\item{betas.Gy}{The betas in the paper: the marginal effects of SNPs on the phenotype (Gy)}

\item{N.Gy}{The sample size of Gy}

\item{Gl}{The reference panel (Gl), such as 1000 Genome}

\item{Z}{The Z matrix in the paper: the marginal/conditional effects of SNPs on the exposures (Gx)}

\item{ridgeTerm}{ridgeTerm = TRUE when the matrix L is singular. Matrix L is obtained from the cholesky decomposition of G0'G0. Default as FALSE.}
}
\description{
The hJAM_egger function is to get the results from the hJAM model with Egger regression. It is for detecting potential pleiotropy
}
\examples{
data(reference_data)
data(betas.Gy)
data(conditional_Z)
hJAM_egger(betas.Gy = betas.Gy, Gl = Gl, N.Gy = 5000, Z = conditional_Z, ridgeTerm = TRUE)
}
\author{
Lai Jiang
}
