% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hJAM_gprior.R
\name{hJAM_gprior}
\alias{hJAM_gprior}
\title{Fit hJAM with putting gprior on the parameter (a Bayesian approach)}
\usage{
hJAM_gprior(betas.Gy, N.Gy, Gl, A, a_sigma = 1, b_sigma = 9,
  trait.variance = 1, ridgeTerm = FALSE)
}
\arguments{
\item{betas.Gy}{The betas in the paper: the marginal effects of SNPs on the phenotype (Gy)}

\item{N.Gy}{The sample size of Gy}

\item{Gl}{The reference panel (Gl), such as 1000 Genome}

\item{A}{The A matrix in the paper: the marginal/conditional effects of SNPs on the exposures (Gx)}

\item{a_sigma}{The scale parameter of the prior, default = 1}

\item{b_sigma}{The scale parameter of the prior, default = 9}

\item{trait.variance}{The variance of the trait (outcome), default = 1}

\item{ridgeTerm}{ridgeTerm = TRUE when the matrix L is singular. Matrix L is obtained from the cholesky decomposition of G0'G0. Default as FALSE.}
}
\description{
The hJAM function is to get the results from the hJAM model using input data
}
\examples{
data(Gl)
data(betas.Gy)
data(conditional_A)
hJAM_gprior(betas.Gy, Gl, N.Gy = 459324, A = conditional_A,a_sigma = 1,
b_sigma = 9, trait.variance = 1, ridgeTerm = TRUE)
}
\author{
Lai Jiang
}
