% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAM_A.R
\name{JAM_A}
\alias{JAM_A}
\title{Compute conditional A matrix}
\usage{
JAM_A(marginalA, Geno, N.Gx, eaf_Gx = NULL, ridgeTerm = TRUE)
}
\arguments{
\item{marginalA}{the marginal effects of SNPs on the exposures (Gx).}

\item{Geno}{the reference panel (Geno), such as 1000 Genome}

\item{N.Gx}{the sample size of each Gx. It can be a scalar or a vector. If there are multiple X's from different Gx, it should be a vector including the sample size of each Gx. If all alphas are from the same Gx, it could be a scalar.}

\item{eaf_Gx}{the effect allele frequency of the SNPs in the Gx data.}

\item{ridgeTerm}{ridgeTerm = TRUE when the matrix L is singular. Matrix L is obtained from the cholesky decomposition of G0'G0. Default as TRUE.}
}
\value{
A matrix with conditional estimates which are converted from marginal estimates using the JAM model.
}
\description{
The \code{JAM_A} function is to get the conditional A matrix by using marginal A matrix
}
\examples{
data(MI)
JAM_A(marginalA = MI.marginal.Amatrix, Geno = MI.Geno, N.Gx = c(339224, 659316), ridgeTerm = TRUE)
JAM_A(marginalA = MI.marginal.Amatrix, Geno = MI.Geno, N.Gx = c(339224, 659316),
eaf_Gx = MI.SNPs_info$ref_frq)
}
\author{
Lai Jiang
}
