% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAM_A.R
\name{JAM_alphas}
\alias{JAM_alphas}
\title{Compute conditional alphas}
\usage{
JAM_alphas(alphas, Geno, N.Gx, eaf_Gx = NULL, ridgeTerm = FALSE)
}
\arguments{
\item{alphas}{the marginal effects of SNPs on one exposure (Gx).}

\item{Geno}{the reference panel (Geno), such as 1000 Genome}

\item{N.Gx}{the sample size of the Gx. It can be a scalar.}

\item{eaf_Gx}{TBD}

\item{ridgeTerm}{ridgeTerm = TRUE when the matrix L is singular. Matrix L is obtained from the cholesky decomposition of G0'G0. Default as FALSE}
}
\value{
A vector with conditional estimates which are converted from marginal estimates using the JAM model.
}
\description{
The JAM_alphas function is to compute the conditional alpha vector for each X
If only one X in the model, please use JAM_alphas instead of JAM_A
A sub-step in the JAM_A function
}
\examples{
data(MI)
JAM_alphas(alphas = MI.marginal.Amatrix[, 1], Geno = MI.Geno, N.Gx = 339224, ridgeTerm = TRUE)
JAM_alphas(alphas = MI.marginal.Amatrix[, 1], Geno = MI.Geno, N.Gx = 339224, eaf_Gx = MI.SNPs_info$ref_frq, ridgeTerm = TRUE)
}
\references{
Lai Jiang, Shujing Xu, Nicholas Mancuso, Paul J. Newcombe, David V. Conti (2020).
A Hierarchical Approach Using Marginal Summary Statistics for Multiple Intermediates
in a Mendelian Randomization or Transcriptome Analysis. \emph{bioRxiv}
\url{https://doi.org/10.1101/2020.02.03.924241}.
}
\author{
Lai Jiang
}
