% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mJAM_helper.R
\name{LogisticToLinearEffects}
\alias{LogisticToLinearEffects}
\title{Transform log odds ratios to linear effects}
\usage{
LogisticToLinearEffects(
  log.ors = NULL,
  log.or.ses = NULL,
  snp.genotype.sds = NULL,
  mafs = NULL,
  n = NULL,
  p.cases = NULL
)
}
\arguments{
\item{log.ors}{A vector of log odds ratios}

\item{log.or.ses}{A vector of the standard errors of the log ORs}

\item{snp.genotype.sds}{A vector of standard deviations of genotypes (optional if `mafs` is provided)}

\item{mafs}{A vector of effective allele frequencies (optional if `snp.genotype.sds` is provided)}

\item{n}{Sample size in the GWAS where we obtained `log.ors`}

\item{p.cases}{A numeric value of the proportion of cases in the GWAS.}
}
\value{
Transformed linear effect estimates, and transformed standards errors of linear effects.
}
\description{
Adopted from R2BGLiMS::JAM_LogisticToLinearEffects. Reference: Benner 2015, FINEMAP
}
