% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marginal_A}
\alias{marginal_A}
\title{Example marginal A matrix of hJAM}
\format{An object of class \code{matrix} with 19 rows and 2 columns.}
\usage{
marginal_A
}
\description{
The marginal A matrix for the examples in hJAM package
}
\keyword{datasets}
