% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marginal_Z}
\alias{marginal_Z}
\title{Example marginal Z matrix of hJAM}
\format{An object of class \code{matrix} with 19 rows and 2 columns.}
\usage{
marginal_Z
}
\description{
The marginal Z matrix for the examples in hJAM package
}
\keyword{datasets}
