% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{MI}
\alias{MI}
\alias{MI.marginal.Amatrix}
\alias{MI.Amatrix}
\alias{MI.Geno}
\alias{MI.betas.gwas}
\alias{MI.SNPs_info}
\title{Example data of hJAM}
\format{
The \code{MI} object is a set of data sets which was used to estimate the causal effect of body mass index and type 2 diabetes on the risk of myocardial infarction.
\describe{
   \item{MI.marginal.Amatrix}{The marginal \eqn{\hat{A}}{A hat} matrix. Column one and two are the marginal estimates of the SNPs on body mass index from GIANT consortium (n = 339,224) (Locke et al., 2015) and type 2 diabetes from DIAGRAM+GERA+UKB (n = 659,316) (Xue et al., 2018), respectively}
   \item{MI.Amatrix}{The conditional \eqn{\hat{A}}{A hat} matrix composed by JAM and the marginal \eqn{\hat{A}}{A hat} matrix. Column one and two are the conditional effect estimates of the SNPs on body mass index and type 2 diabetes, respectively.}
   \item{MI.Geno}{The reference genotype data from the European-ancestry population in 1000 Genome Project (Consortium, 2015).}
   \item{MI.betas.gwas}{The b vector. The association estimates between selected SNPs and the risk of myocardial infarction from UK Biobank (Sudlow et al., 2015).}
   \item{MI.SNPs_info}{The SNP information. Five columns included: the RSID, reference allele, reference allele frequency, if BMI significant and if T2D significant. The last two columns are indicator variables for the SNPs which are genome-wide significant associated with BMI/T2D.}
   }
}
\description{
Real data for BMI/T2D on the risk of Myocardial infarction
}
\references{
Consortium GP. A global reference for human genetic variation. Nature 2015; 526: 68.

Locke, Adam E., et al. Genetic studies of body mass index yield new insights for obesity biology. Nature 518.7538 (2015): 197-206.

Xue, Angli, et al. Genome-wide association analyses identify 143 risk variants and putative regulatory mechanisms for type 2 diabetes. Nature communications 9.1 (2018): 1-14.

Sudlow, Cathie, et al. UK biobank: an open access resource for identifying the causes of a wide range of complex diseases of middle and old age. Plos med 12.3 (2015): e1001779.
}
