% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mJAM_get_condp_logp.R
\name{mJAM_get_condp}
\alias{mJAM_get_condp}
\title{Get conditional p-value under mJAM model}
\usage{
mJAM_get_condp(
  GItGI,
  GIty,
  yty,
  yty_med,
  N_GWAS,
  g = NULL,
  selected_id,
  use_median_yty_ethnic = NULL,
  rare_id = NULL
)
}
\arguments{
\item{GItGI}{A list of transformed statistics from `get_XtX()` for each study.}

\item{GIty}{A list of transformed statistics from `get_z()` for each study.}

\item{yty}{A list of transformed statistics from `get_yty()` for each study.}

\item{yty_med}{A numeric vector of median yty across all SNPs within each study.}

\item{N_GWAS}{A numeric vector of GWAS sample size for each study.}

\item{g}{Hyperparameter in g-prior. If `NULL`, it will be set to `sum(N_GWAS)`.}

\item{selected_id}{A numeric vector of IDs of previously selected index SNP(s).}

\item{use_median_yty_ethnic}{A numeric vector of study index in which median_yty is used for all SNPs in `selected_id`.}

\item{rare_id}{A numeric vector of IDs for rare SNP(s) which we do not apply weighting. Instead, we use the individual estimate of yty for these SNPs for robustness.}
}
\value{
\describe{
   \item{which_condp_min}{The index of which SNP has the smallest conditional p-value.}
   \item{condp_min}{The smallest conditional p-value.}
   \item{condp}{A vector of all conditional p-values.}
   \item{effect_est}{A vector of all conditional effect estimates.}
   \item{se_est}{A vector of standard errors of all the conditional effect estimates.}
   \item{condp_mx}{A complete matrix recording all conditional effect est & se for testing SNPs and `selected_id`.}
}
}
\description{
Get conditional p-value under mJAM model
}
\author{
Jiayi Shen
}
