% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mJAM_get_condp_logp.R
\name{mJAM_get_condp_selected}
\alias{mJAM_get_condp_selected}
\title{Get conditional p-value for selected (index SNPs) under mJAM model}
\usage{
mJAM_get_condp_selected(
  GItGI,
  GIty,
  yty,
  yty_med,
  N_GWAS,
  g = NULL,
  selected_id,
  use_robust_var_est = FALSE,
  use_median_yty_ethnic = NULL,
  rare_SNPs = NULL
)
}
\arguments{
\item{GItGI}{A list of transformed statistics from `get_XtX()` for each study.}

\item{GIty}{A list of transformed statistics from `get_z()` for each study.}

\item{yty}{A list of transformed statistics from `get_yty()` for each study.}

\item{yty_med}{A numeric vector of median yty across all SNPs within each study.}

\item{N_GWAS}{A numeric vector of GWAS sample size for each study.}

\item{g}{Hyperparameter in g-prior. If `NULL`, it will be set to `sum(N_GWAS)`.}

\item{selected_id}{A numeric vector of IDs of previously selected index SNP(s).}

\item{use_robust_var_est}{whether to use linear combination of median yty and individual yty.}

\item{use_median_yty_ethnic}{A numeric vector of study index in which median_yty is used for all SNPs in `selected_id`.}

\item{rare_SNPs}{A character vector for rare SNP(s) which we do not apply weighting. Instead, we use the individual estimate of yty for these SNPs for robustness.}
}
\value{
\describe{
   \item{b_joint}{The estimated conditional effect size when all SNPs in `selected_id` are in one mJAM model.}
   \item{b_joint_var}{The variance of `b_joint`.}
   \item{condp}{A vector of all conditional p-values for `b_joint`.}
}
}
\description{
Get conditional p-value for selected (index SNPs) under mJAM model
}
\author{
Jiayi Shen
}
