% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mJAM_build_CS.R
\name{mJAM_get_PrM}
\alias{mJAM_get_PrM}
\title{Get Pr(Model) based on BF-type model probability}
\usage{
mJAM_get_PrM(
  GItGI,
  GIty,
  yty,
  yty_med,
  N_GWAS,
  C_id,
  prev_X_list = NULL,
  g = NULL,
  rare_SNPs = NULL,
  use_robust_var_est = FALSE
)
}
\arguments{
\item{GItGI}{A list of transformed statistics from `get_XtX()` for each study.}

\item{GIty}{A list of transformed statistics from `get_z()` for each study.}

\item{yty}{A list of transformed statistics from `get_yty()` for each study.}

\item{yty_med}{A numeric vector of median yty across all SNPs within each study.}

\item{N_GWAS}{A numeric vector of GWAS sample size for each study.}

\item{C_id}{An ingeter vector of IDs for the SNPs to be tested.}

\item{prev_X_list}{A numeric vector of the ID(s) of previously selected index SNP(s).}

\item{g}{The pre-specified `g` in `g`-prior formulation.}

\item{rare_SNPs}{A numeric vector of ID(s) for rare SNP(s) which we do not apply weighting. Instead, we use the individual estimate of yty for these SNPs for robustness.}

\item{use_robust_var_est}{whether to use linear combination of median yty and individual yty.}
}
\value{
\describe{
   \item{post_prob}{Posterior Pr(Model) for each SNPs in `C_id`.}
   \item{R2_est}{R2 estimates of every one-SNP model (one for each SNPs in `C_id`).}
   \item{n_miss}{An integer vector of how many studies have missing values for each SNP.}
}
}
\description{
Also apply weighting to get robust estimates of yty
}
\author{
Jiayi Shen
}
