% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mJAM_build_CS.R
\name{mJAM_get_PrMed}
\alias{mJAM_get_PrMed}
\title{Get Pr(Mediation) based on causal mediation models}
\usage{
mJAM_get_PrMed(
  GItGI,
  GIty,
  yty,
  yty_med,
  N_GWAS,
  g = NULL,
  C_id,
  X_id,
  prev_X_list
)
}
\arguments{
\item{GItGI}{A list of transformed statistics from `get_XtX()` for each study.}

\item{GIty}{A list of transformed statistics from `get_z()` for each study.}

\item{yty}{A list of transformed statistics from `get_yty()` for each study.}

\item{yty_med}{A numeric vector of median yty across all SNPs within each study.}

\item{N_GWAS}{A numeric vector of GWAS sample size for each study.}

\item{g}{The pre-specified `g` in `g`-prior formulation.}

\item{C_id}{An ingeter vector of IDs for the SNPs to be tested.}

\item{X_id}{An integer specifying the ID of the index SNP.}

\item{prev_X_list}{A numeric vector of the ID(s) of previously selected index SNP(s).}
}
\value{
\describe{
   \item{Post_Med_Prob}{Posterior Pr(Mediation) for each SNPs in C_id.}
   \item{Med_Effect_Size}{Posterior mediation effect size for each SNPs in C_id.}
   \item{Med_var_CX}{Posterior variance of mediation effect in models with both C and X.}
   \item{Med_var_C}{Posterior variance of mediation effect in models with C only.}
}
}
\description{
Also apply weighting to get robust estimates of yty
}
\author{
Jiayi Shen
}
