% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mJAM_LDpruning.R
\name{mJAM_LDpruning}
\alias{mJAM_LDpruning}
\title{Pruning SNPs based on LD}
\usage{
mJAM_LDpruning(target, testing, R, within_thre = 0.95, across_thre = 0.8)
}
\arguments{
\item{target}{Target SNP ID.}

\item{testing}{IDs of SNPs to be tested.}

\item{R}{a list of correlation matrix of all SNPs.}

\item{within_thre}{threshold of absolute correlation with selected index SNP(s) within a single population. If a SNP's correlation with any selected index SNP is greater than this threshold in at least one population, it will be excluded from subsequent rounds of index SNP selection.}

\item{across_thre}{threshold of absolute correlation with selected index SNP(s) across all populations. If a SNP's correlation with any selected index SNP is greater than this threshold in all populations, it will be excluded from subsequent rounds of index SNP selection.}
}
\value{
\describe{
   \item{remove_within}{SNP IDs to be pruned due to high within-population correlation}
   \item{remove_across}{SNP IDs to be pruned due to high across-population correlation}
}
}
\description{
Pruning SNPs based on LD
}
\author{
Jiayi Shen
}
