% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mJAM_SuSiE.R
\name{mJAM_SuSiE}
\alias{mJAM_SuSiE}
\title{Run mJAM with SuSiE}
\usage{
mJAM_SuSiE(
  Marg_Result = NULL,
  EAF_Result = NULL,
  N_GWAS,
  X_ref,
  filter_rare = FALSE,
  rare_freq = NULL,
  SuSiE_num_comp = 10,
  SuSiE_coverage = 0.95,
  SuSiE_min_abs_corr = 0.5,
  max_iter = 500,
  estimate_residual_variance = F
)
}
\arguments{
\item{Marg_Result}{A data frame with marginal summary statistics from all studies. Col1: SNP name; Col2: Effect sizes from study #1; Col3: Std Errors of effect sizes from study #1; ...}

\item{EAF_Result}{A data frame with effect allele frequency (EAF) from all studies. Col1: SNP name; Col2: EAF from study #1; Col3: EAF from study #2; ...}

\item{N_GWAS}{A vector of sample sizes in all original GWAS studies.}

\item{X_ref}{A list of matrices with individual-level SNP dosage data in each study/population.}

\item{filter_rare}{A logical variable indicating whether to filter rare SNPs before the analysis. Default is `FALSE.` If `TRUE`, then please specify `rare_freq`.}

\item{rare_freq}{A vector of frequencies between 0 and 0.5 to specify the minor allele frequency cut-off if you want to filter rare SNPs before the analysis. Please also set `filter_rare` to be TRUE. For example, if there are 3 populations, then rare_freq = c(0.01, 0, 0.01) means SNPs with MAF < 0.01 in pop 1 and MAF < 0.01 in pop 3 will be removed from analysis.}

\item{SuSiE_num_comp}{SuSiE argument. The maximum number of causal SNPs that you want to select. Default is 10.}

\item{SuSiE_coverage}{SuSiE argument. The required coverage of credible sets. Default is 0.95.}

\item{SuSiE_min_abs_corr}{SuSiE argument. Minimum absolute correlation allowed in a credible set.}

\item{max_iter}{SuSiE argument. Maximum iterations to perform.}

\item{estimate_residual_variance}{SuSiE argument. If `TRUE`, then the susie algorithm is updating residual variance estimate during iterations. If `FALSE`, then use the residual variance is a fixed value, which is usually var(Y).}
}
\value{
\describe{
   \item{summary}{A table of the SuSiE posterior inclusion probabilities (PIPs), posterior mean, and posterior sd of all SNPs.}
   \item{fit}{SuSiE fit object.}
}
}
\description{
fitting mJAM-SuSiE
}
\author{
Jiayi Shen
}
