% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{output.format}
\alias{output.format}
\title{Keep the output as three digits}
\usage{
output.format(x, ...)
}
\arguments{
\item{x}{input}

\item{...}{other options you want to put in}
}
\description{
Keep the output as three digits
}
\author{
Lai Jiang
}
