% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ENhJAM}
\alias{print.ENhJAM}
\title{Print out for EN-hJAM}
\usage{
\method{print}{ENhJAM}(x, ...)
}
\arguments{
\item{x}{obejct output from ENhJAM}

\item{...}{other options you want to put in}
}
\description{
Print out for EN-hJAM
}
\author{
Lai Jiang
}
