% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.hJAM}
\alias{print.hJAM}
\title{Print out for hJAM}
\usage{
\method{print}{hJAM}(x, ...)
}
\arguments{
\item{x}{object output by hJAM}

\item{...}{other options you want to put in}
}
\description{
Print out for hJAM_lnreg
}
\author{
Lai Jiang
}
