% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.hJAM_egger}
\alias{print.hJAM_egger}
\title{Print out for hJAM_egger}
\usage{
\method{print}{hJAM_egger}(x, ...)
}
\arguments{
\item{x}{input}

\item{...}{other options you want to put in}
}
\description{
Print out for hJAM_egger
}
\author{
Lai Jiang
}
