% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.hJAM_gprior}
\alias{print.hJAM_gprior}
\title{Print out for hJAM_gprior}
\usage{
\method{print}{hJAM_gprior}(x, ...)
}
\arguments{
\item{x}{input}

\item{...}{other options you want to put in}
}
\description{
Print out for hJAM_lnreg
}
\author{
Lai Jiang
}
