% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.hJAM_lnreg}
\alias{print.hJAM_lnreg}
\title{Print out for hJAM_lnreg}
\usage{
\method{print}{hJAM_lnreg}(x, ...)
}
\arguments{
\item{x}{input}

\item{...}{other options you want to put in}
}
\description{
Print out for hJAM_lnreg
}
\author{
Lai Jiang
}
