% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.SHAJAM}
\alias{print.SHAJAM}
\title{Print out for SHA-JAM}
\usage{
\method{print}{SHAJAM}(x, ...)
}
\arguments{
\item{x}{obejct output from SHAJAM}

\item{...}{other options you want to put in}
}
\description{
Print out for SHA-JAM
}
\author{
Lai Jiang
}
