% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reghJAM.R
\name{reghJAM}
\alias{reghJAM}
\title{Regularized hJAM}
\usage{
reghJAM(
  betas.Gy,
  betas_se.Gy = NULL,
  N.Gy,
  raf.Gy = NULL,
  Geno,
  A,
  selection_alg = "en",
  ridgeTerm = FALSE
)
}
\arguments{
\item{betas.Gy}{The betas in the paper: the marginal effects of SNPs on the phenotype (Gy)}

\item{betas_se.Gy}{The standard errors of the betas}

\item{N.Gy}{The sample size of the GWAS where you obtain the betas.Gy and betas_se.Gy}

\item{raf.Gy}{The reference allele frequency of the SNPs in betas.Gy}

\item{Geno}{The individual level data of the reference panel. Must have the same order of SNPs as in the betas.Gy.}

\item{A}{The conditional A matrix.}

\item{selection_alg}{The selection algorithm. It can be SHA-JAM ("susie") or regularized hJAM ("en" for elastic net and "lasso" for LASSO).}

\item{ridgeTerm}{Add a small elelment to the diagnoal of X'X to make the matrix invertable.}
}
\value{
An object of the Regularized hJAM

\describe{
   \item{Exposure}{The intermediates, such as the modifiable risk factors in Mendelian Randomization and gene expression in transcriptome analysis.}
   \item{numSNP}{The number of SNPs that the user use in the instrument set.}
   \item{Estimate}{The conditional estimates of the associations between intermediates and the outcome.}
   \item{StdErr}{The standard error of the conditional estimates of the associations between intermediates and the outcome.}
   \item{Lower.CI}{The lower bound of the 95\% confidence interval of the estimates.}
   \item{Upper.CI}{The upper bound of the 95\% confidence interval of the estimates.}
   \item{Pvalue}{The p value of the estimates with a type-I error equals 0.05.}
}
}
\description{
Function to implement regularized hJAM, including elastic net hJAM and lasso hJAM.
}
\author{
Lai Jiang
}
