% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHAJAM.R
\name{SHAJAM}
\alias{SHAJAM}
\title{SHA-JAM
Fit SHA-JAM}
\usage{
SHAJAM(
  betas.Gy,
  betas_se.Gy = NULL,
  N.Gy,
  raf.Gy = NULL,
  Geno,
  A,
  L.susie = NULL,
  min_abs_corr = NULL,
  coverage = 0.95,
  estimate_residual_variance = TRUE,
  max_iter = 500
)
}
\arguments{
\item{betas.Gy}{The betas in the paper: the marginal effects of SNPs on the phenotype (Gy)}

\item{betas_se.Gy}{The standard errors of the betas}

\item{N.Gy}{The sample size of the GWAS where you obtain the betas.Gy and betas_se.Gy}

\item{raf.Gy}{The reference allele frequency of the SNPs in betas.Gy}

\item{Geno}{The individual level data of the reference panel. Must have the same order of SNPs as in the betas.Gy.}

\item{A}{The conditional A matrix.}

\item{L.susie}{The largest number of credible set allowed in SHA-JAM. Required by SHA-JAM.}

\item{min_abs_corr}{The requested minimum absolute correlation coefficient between intermediates within one credible set. Required by SHA-JAM.}

\item{coverage}{The coverage of credible set. Default is 0.95. Required by SHA-JAM.}

\item{estimate_residual_variance}{If estimate the residual variance in the fitting procedure of SHA-JAM. Default as TRUE. Required by SHA-JAM.}

\item{max_iter}{The number of maximum iterations in fitting SHA-JAM. Required by SHA-JAM.}
}
\value{
An object of the SHAJAM

\describe{
   \item{Exposure}{The intermediates, such as the modifiable risk factors in Mendelian Randomization and gene expression in transcriptome analysis.}
   \item{numSNP}{The number of SNPs that the user use in the instrument set.}
   \item{Estimate}{The conditional estimates of the associations between intermediates and the outcome.}
   \item{StdErr}{The standard error of the conditional estimates of the associations between intermediates and the outcome.}
   \item{Lower.CI}{The lower bound of the 95\% confidence interval of the estimates.}
   \item{Upper.CI}{The upper bound of the 95\% confidence interval of the estimates.}
   \item{Pvalue}{The p value of the estimates with a type-I error equals 0.05.}
}
}
\description{
Function to implement SHA-JAM
}
\author{
Lai Jiang
}
