% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{SNPs_heatmap}
\alias{SNPs_heatmap}
\title{Heatmap for all the SNPs used in the analysis}
\usage{
SNPs_heatmap(Geno, show.variables = FALSE)
}
\arguments{
\item{Geno}{The reference panel (Geno) of the SNPs that the user use in the analysis, such as 1000 Genome}

\item{show.variables}{Select to show the variables name or not. Default set to be FALSE.}
}
\description{
To generate the heatmap of all the SNPs that the user use in the analysis
}
\examples{
data(MI.Rdata)
SNPs_heatmap(Geno = MI.Geno[, 1: 10])
SNPs_heatmap(Geno = MI.Geno[, 1: 10], show.variable = TRUE)
}
\author{
Lai Jiang
}
