% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{SNPs_scatter_plot}
\alias{SNPs_scatter_plot}
\title{Scatter plot for SNPs vs. one intermediate in the analysis}
\usage{
SNPs_scatter_plot(alphas, betas.Gy, X.label = NULL)
}
\arguments{
\item{alphas}{The effects of SNPs on the intermediate (i.e. exposure/risk factor) (Gx).}

\item{betas.Gy}{The betas in the paper: the marginal effects of SNPs on the phenotype (Gy)}

\item{X.label}{The label of the intermediate (i.e. exposure/risk factor). Default is NULL.}
}
\value{
A set of scatter plots with x-axis being the conditional \eqn{\alpha}{alpha} estimates for each
intermediate and y-axis being the \eqn{\beta}{beta} estimates.
}
\description{
To generate the scatter plot of the SNPs vs. one intermediate that the user use in the analysis
}
\examples{
data(MI)
SNPs_scatter_plot(alphas = MI.Amatrix[, 1], betas.Gy = MI.betas.gwas, X.label = "BMI")
}
\author{
Lai Jiang
}
