% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susieJAM.R
\name{susieJAM_A}
\alias{susieJAM_A}
\title{Compute conditional A using SuSiE JAM}
\usage{
susieJAM_A(
  marginalA,
  marginalA_se,
  N.Gx,
  raf.Gy = NULL,
  Geno,
  inclusion.indicator,
  L.cs,
  min_abs_corr,
  max_iter,
  coverage,
  estimate_residual_variance = TRUE
)
}
\arguments{
\item{marginalA}{the marginal effects of SNPs on the exposures (Gx).}

\item{marginalA_se}{the standard error of the marginal effects of SNPs on the exposures (Gx).}

\item{N.Gx}{the sample size of each Gx. It can be a scalar or a vector. If there are multiple X's from different Gx, it should be a vector including the sample size of each Gx. If all alphas are from the same Gx, it could be a scalar.}

\item{raf.Gy}{the effect allele frequency of the SNPs in the Gx data.}

\item{Geno}{the reference panel (Geno), such as 1000 Genome。}

\item{inclusion.indicator}{The matrix of inclusion indicator of SNPs for each intermediate. Included as 1; otherwise 0.}

\item{L.cs}{A susie input parameter. Number of components (nonzero elements) in the SuSiE regression model. If L.cs is larger than the number of covariate (p), L.cs is set to p.}

\item{min_abs_corr}{A susie input parameter. Minimum of absolute value of correlation allowed in a credible set. The default, 0.5, corresponds to squared correlation of 0.25, which is a commonly used threshold for genotype data in genetics studies.}

\item{max_iter}{Maximum number of iterations in SuSiE fitting.}

\item{coverage}{Default as 0.95.The coveralge level of the credible set.}

\item{estimate_residual_variance}{Default as TRUE. Estimate the residual variance in each iteration of SuSiE fitting.}
}
\value{
A matrix with conditional estimates which are converted from marginal estimates using the susie JAM model.
}
\description{
The \code{susieJAM_A} function is to get the conditional A matrix by using marginal A matrix
}
\author{
Lai Jiang
}
