% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susieJAM.R
\name{susieJAM_alphas}
\alias{susieJAM_alphas}
\title{Compute conditional alphas using SuSiE JAM}
\usage{
susieJAM_alphas(
  marginalA,
  marginalA_se,
  N.Gx,
  raf.Gy = NULL,
  Geno,
  L.cs = 10,
  min_abs_corr = 0.6,
  max_iter = 100,
  coverage = 0.95,
  estimate_residual_variance = FALSE
)
}
\arguments{
\item{marginalA}{the marginal effects of SNPs on one exposure (Gx).}

\item{marginalA_se}{the standard error of the marginal effects of SNPs on one outcome (Gx).}

\item{N.Gx}{the sample size of the Gx. It can be a scalar.}

\item{raf.Gy}{The vector of the minor allele frequency or effect allele frequency in the GWAS.}

\item{Geno}{the reference panel (Geno), such as 1000 Genome. The reference data has to be centered.}

\item{L.cs}{A susie input parameter. Number of components (nonzero elements) in the SuSiE regression model. If L.cs is larger than the number of covariate (p), L.cs is set to p.}

\item{min_abs_corr}{A susie input parameter. Minimum of absolute value of correlation allowed in a credible set. The default, 0.5, corresponds to squared correlation of 0.25, which is a commonly used threshold for genotype data in genetics studies.}

\item{max_iter}{Maximum number of iterations in SuSiE fitting.}

\item{coverage}{Default as 0.95.The coveralge level of the credible set.}

\item{estimate_residual_variance}{Default as TRUE. Estimate the residual variance in each iteration of SuSiE fitting.}
}
\description{
The \code{susieJAM_alphas} function is to perform the variable selection and compute the selected conditional alpha vector for one intermediate.
If only one intermediate in the model, please use susieJAM_alphas instead of susieJAM_A
}
\examples{
data(GTEx.PrCa)
include.SNPs = which(GTEx.PrCa.inclusion.indicator[,1]==1)
susieJAM_alphas(marginalA = GTEx.PrCa.marginal.A[include.SNPs, 1],
marginalA_se = GTEx.PrCa.marginal.A.se[include.SNPs, 1], raf.Gy = GTEx.PrCa.maf.gwas[include.SNPs],
Geno = GTEx.PrCa.Geno[, include.SNPs], N.Gx = 620, L.cs = 10, min_abs_corr = 0.5)
}
\author{
Lai Jiang
}
