% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cond_alpha.R
\name{get_cond_alpha}
\alias{get_cond_alpha}
\title{Compute conditional alphas}
\usage{
get_cond_alpha(alphas, Gl, N.Gx, ridgeTerm = FALSE)
}
\arguments{
\item{alphas}{the marginal effects of SNPs on one exposure (Gx).}

\item{Gl}{the reference panel (Gl), such as 1000 Genome}

\item{N.Gx}{the sample size of the Gx. It can be a scalar.}

\item{ridgeTerm}{ridgeTerm = TRUE when the matrix L is singular. Matrix L is obtained from the cholesky decomposition of G0'G0. Default as FALSE}
}
\description{
The get_cond_alpha function is to compute the conditional alpha vector for each X
If only one X in the model, please use get_cond_alpha instead of get_cond_Z
A sub-step in the get_cond_Z function
}
\examples{
data(reference_data)
data(betas.Gy)
data(marginal_Z)
get_cond_alpha(alphas = marginal_Z[, 1], Gl = Gl, N.Gx = 1000, ridgeTerm = TRUE)
}
\author{
Lai Jiang
}
