% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Gl}
\alias{Gl}
\title{Example reference data of hJAM}
\format{The \code{Gl} object is a data matrix with 2467 individual of 210 SNPs from 1000 Genome project.}
\usage{
Gl
}
\description{
The real data example from hJAM paper
}
\references{
Consortium GP. A global reference for human genetic variation. Nature 2015; 526: 68.
}
\keyword{datasets}
