% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hJAM_egger.R
\name{hJAM_egger}
\alias{hJAM_egger}
\title{Fit hJAM with Egger regression}
\usage{
hJAM_egger(betas.Gy, N.Gy, Gl, A, ridgeTerm = FALSE)
}
\arguments{
\item{betas.Gy}{The betas in the paper: the marginal effects of SNPs on the phenotype (Gy)}

\item{N.Gy}{The sample size of Gy}

\item{Gl}{The reference panel (Gl), such as 1000 Genome}

\item{A}{The A matrix in the paper: the marginal/conditional effects of SNPs on the exposures (Gx)}

\item{ridgeTerm}{ridgeTerm = TRUE when the matrix L is singular. Matrix L is obtained from the cholesky decomposition of G0'G0. Default as FALSE.}
}
\description{
The hJAM_egger function is to get the results from the hJAM model with Egger regression. It is for detecting potential pleiotropy
}
\examples{
data(Gl)
data(betas.Gy)
data(conditional_A)
hJAM_egger(betas.Gy = betas.Gy, Gl = Gl, N.Gy = 459324, A = conditional_A, ridgeTerm = TRUE)
}
\references{
Lai Jiang, Shujing Xu, Nicholas Mancuso, Paul J. Newcombe, David V. Conti (2020).
A Hierarchical Approach Using Marginal Summary Statistics for Multiple Intermediates
in a Mendelian Randomization or Transcriptome Analysis. \emph{bioRxiv}
\url{https://doi.org/10.1101/2020.02.03.924241}.
}
\author{
Lai Jiang
}
