% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mJAM_SuSiE.R
\name{mJAM_SuSiE}
\alias{mJAM_SuSiE}
\title{Run mJAM with SuSiE}
\usage{
mJAM_SuSiE(
  marginal.betas = NULL,
  marginal.se = NULL,
  EAFs,
  N_GWAS,
  X_ref,
  SNP_names = NULL,
  logORToBeta = FALSE,
  p_cases = NULL,
  replace_missing = TRUE,
  SuSiE_num_comp = 10,
  SuSiE_coverage = 0.95,
  SuSiE_min_abs_corr = 0.5,
  max_iter = 500,
  estimate_residual_variance = F
)
}
\arguments{
\item{marginal.betas}{A list of marginal effect estimates from each study/population. The length of the list should be equal to the number of studies.}

\item{marginal.se}{A list of marginal standard error of `marginal.betas` from each study/population.}

\item{EAFs}{A list of effect allele frequencies from each study/population.}

\item{N_GWAS}{A vector of sample sizes in all original GWAS studies.}

\item{X_ref}{A list of matrices with individual-level SNP dosage data in each study/population.}

\item{SNP_names}{A character vector of the names of the SNPs. Note that the order of SNP names should be the same as `EAFs`, `X_ref`, `marginal.betas` and `marginal.se`.}

\item{logORToBeta}{If `TRUE`, transform the marginal logOR estimates in `marginal.betas` to marginal linear effects.}

\item{p_cases}{(Required if `logORToBeta == TRUE`) The proportion of cases in each study/population, use `c()` to combine.}

\item{replace_missing}{If `TRUE`, then replace missing SNPs with zeros in the summary statistics of the missing study(s). Default is TRUE.}

\item{SuSiE_num_comp}{SuSiE argument. The maximum number of causal SNPs that you want to select. Default is 10.}

\item{SuSiE_coverage}{SuSiE argument. The required coverage of credible sets. Default is 0.95.}

\item{SuSiE_min_abs_corr}{SuSiE argument. Minimum absolute correlation allowed in a credible set.}

\item{max_iter}{SuSiE argument. Maximum iterations to perform.}

\item{estimate_residual_variance}{SuSiE argument. If `TRUE`, then the susie algorithm is updating residual variance estimate during iterations. If `FALSE`, then use the residual variance is a fixed value, which is usually var(Y).}
}
\value{
\describe{
   \item{summary}{A table of the SuSiE posterior inclusion probabilities (PIPs), posterior mean, and posterior sd of all SNPs.}
   \item{fit}{SuSiE fit object.}
}
}
\description{
fitting mJAM-SuSiE
}
\author{
Jiayi Shen
}
