% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mJAM_SuSiE_helper.R
\name{mJAM_SuSiE_get_cs}
\alias{mJAM_SuSiE_get_cs}
\title{Get and tidy SuSiE credible sets}
\usage{
mJAM_SuSiE_get_cs(susie_fit, SNP_names, coverage = 0.95)
}
\arguments{
\item{susie_fit}{A SuSiE fit object}

\item{SNP_names}{A character vector of the names of the SNPs. Note that the order of SNP names should be the same as `EAFs`, `X_ref`, `marginal.betas` and `marginal.se` in `mJAM_SuSiE()`.}

\item{coverage}{A number between 0 and 1 specifying the “coverage” of the estimated confidence sets.}
}
\value{
A table summary of SuSiE credible sets with the following columns:

#' \describe{
   \item{index}{The label for a distinct credible set.}
   \item{coverage}{The empirical coverage of this credible set.}
   \item{CS_size}{The number of SNPs in total in corresponding credible set.}
   \item{index_SNP_id}{The name of the index SNP (SNP with highest posterior probability) in corresponding credible set.}
   \item{CS_SNP_id}{The names of individual SNPs selected in this credible set.}
}
}
\description{
Get and tidy SuSiE credible sets
}
\author{
Jiayi Shen
}
