% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHAJAM.R
\name{SHAJAM}
\alias{SHAJAM}
\title{SHA-JAM
Fit SHA-JAM}
\usage{
SHAJAM(
  betas.Gy,
  betas_se.Gy = NULL,
  N.Gy,
  eaf.Gy = NULL,
  Geno,
  A,
  L.cs = NULL,
  min_abs_corr = NULL,
  coverage = 0.95,
  estimate_residual_variance = TRUE,
  max_iter = 500
)
}
\arguments{
\item{betas.Gy}{The betas in the paper: the marginal effects of SNPs on the phenotype (Gy)}

\item{betas_se.Gy}{The standard errors of the betas}

\item{N.Gy}{The sample size of the GWAS where you obtain the betas.Gy and betas_se.Gy}

\item{eaf.Gy}{The reference allele frequency of the SNPs in betas.Gy}

\item{Geno}{The individual level data of the reference panel. Must have the same order of SNPs as in the betas.Gy.}

\item{A}{The conditional A matrix.}

\item{L.cs}{The largest number of credible set allowed in SHA-JAM. Required by SHA-JAM.}

\item{min_abs_corr}{The requested minimum absolute correlation coefficient between intermediates within one credible set. Required by SHA-JAM.}

\item{coverage}{The coverage of credible set. Default is 0.95. Required by SHA-JAM.}

\item{estimate_residual_variance}{If estimate the residual variance in the fitting procedure of SHA-JAM. Default as TRUE. Required by SHA-JAM.}

\item{max_iter}{The number of maximum iterations in fitting SHA-JAM. Required by SHA-JAM.}
}
\value{
An object of the SHAJAM

\describe{
   \item{numSNP}{The number of SNPs used in the analysis.}
   \item{numX}{The number of intermediates in the analysis.}
   \item{Selected_variable_length}{The number of selected intermediates, regardless of the credible sets.}
   \item{Selected_variable_name}{The label/name for each selected intermediates.}
   \item{Coefficients}{The coefficients of selected intermediates.}
   \item{Selected_variable_pip}{The posterior inclusion probability of each selected intermediate.}
   \item{num_Credible_sets}{Number of credible sets.}
   \item{all_variables}{The label/name for all candidate intermediates.}
   \item{all_variable_pip}{The posterior inclusion probability of all candidate intermediates.}
   \item{all_variable_coefficient}{The coefficients of all candidate intermediates.}
   \item{cs_purity}{The purity of the credibel set selected.}
}
}
\description{
Function to implement SHA-JAM
}
\author{
Lai Jiang
}
