% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHAJAM.R
\name{SHAJAM}
\alias{SHAJAM}
\title{SHA-JAM
Fit SHA-JAM}
\usage{
SHAJAM(
  betas.Gy,
  betas_se.Gy = NULL,
  N.Gy,
  MAF_summary = NULL,
  Gl,
  A,
  selection_alg = "en",
  L.susie = NULL,
  min_abs_corr = NULL,
  coverage = 0.95,
  estimate_residual_variance = TRUE,
  max_iter = 500,
  ridgeTerm = FALSE
)
}
\arguments{
\item{betas.Gy}{The betas in the paper: the marginal effects of SNPs on the phenotype (Gy)}
}
\value{
An object of the SHAJAM

\describe{
   \item{Exposure}{The intermediates, such as the modifiable risk factors in Mendelian Randomization and gene expression in transcriptome analysis.}
   \item{numSNP}{The number of SNPs that the user use in the instrument set.}
   \item{Estimate}{The conditional estimates of the associations between intermediates and the outcome.}
   \item{StdErr}{The standard error of the conditional estimates of the associations between intermediates and the outcome.}
   \item{Lower.CI}{The lower bound of the 95\% confidence interval of the estimates.}
   \item{Upper.CI}{The upper bound of the 95\% confidence interval of the estimates.}
   \item{Pvalue}{The p value of the estimates with a type-I error equals 0.05.}
}
}
\description{

}
\author{
Lai Jiang
}
