% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPs_heatmap.R
\name{SNPs_heatmap}
\alias{SNPs_heatmap}
\title{Heatmap for all the SNPs used in the analysis}
\usage{
SNPs_heatmap(Gl)
}
\arguments{
\item{Gl}{The reference panel (Gl) of the SNPs that the user use in the analysis, such as 1000 Genome}
}
\description{
To generate the heatmap of all the SNPs that the user use in the analysis
}
\examples{
data(Gl)
t = SNPs_heatmap(Gl = Gl)
t
}
\author{
Lai Jiang
}
